comInterfaces_sconscript: correct case of MathPlayer.py string. (#12310)
Fixes #12281

Math can no longer be read in NVDA with mathPlayer.
a regression introduced by pr #12201
comInterfaces_sconscript instructs comtypes to generate Python files from typelibs/mathPlayer.tlb. Comtypes produces MathPlayer.py (note the uppercase M). However, the sconscript target was hardcoded as "mathPlayer.py" (note the lowercase m). This has always been the case, and has not been a problem as scons and Windows is case insensitive and so Scons thought the target had been correctly created. However, pr #12201 makes use of the target's abspath property to rewrite the file with some extra imports. However, the path is made from the hardcoded target, so it has the lowercase m and therefore when the file is written out again, it has a lowercase m, and mathPres cannot then import MathPlayer with an uppercase M.

Description of how this pull request fixes the issue:
Correct the "mathPlayer.py" string in comInterfaces_sconscript to "MathPlayer.py".