Bug Report #12331
NVDA does not announce the names of the menus when walking with left and right arrows in iTunes. This can potentially be fixed by adding a class which helps accesses the text inside of an iTunes menu box. The class added will attempt to provide that functionality. Further additions add an object of this class to the clsList in chooseNVDAObjectOverlayClasses. A comment was not made on the bug report due to the issue unaffecting the user experience in any way and tight scheduling.