In MS Word, fixed table quick navigation when include layout table option is enabled (#11899)
Summary of the issue:
In MS Word/Outlook browse mode, layout tables (i.e. tables with no border) are skipped when using table quick navigation key (T and shift+T), even if "Include layout tables" options is enabled in Browse mode settings.

Description of how this pull request fixes the issue:
In NVDAObjects\window\winword, I have modified the TableWinWordCollectionQuicknavIterator filter method so that the iterator returns a table if borders are enabled or if layout tables should be included.

I have checked in nvdaHelper\remote\winword.cpp that the presence or the absence of a border was already the criterion that defines if a table is a layout table or not. (so nothing done in this file)

Co-authored-by: buddsean <sean@nvaccess.org>