improve state management in speech.py (#12395)
Summary of the issue:
Using shift to resume speech after using shift to pause speech does not resume speech.

This is because when isPaused is imported through speech\__init__.py it is does not reference the global variable used in speech.py

Description of how this pull request fixes the issue:
Removes the use of top level variables in speech.py in favour of a private dataclass to manage state, and an enum for certain modes. Constants (or rather, variables with names in all capitals) remain as is.