log focusing chrome better during system tests (#12415)
It is hard to know why a certain system test is failing. The test seems to be unable to focus the content of the focused chrome window. Logging information about the focus windows will help compare failing and succeeding tests.

Some further investigation in this PR #12408 led me to think that the incorrect window is gaining focus

Specifically where the "before test sample marker" was unable to be read when cycling focus using F6

Description of how this pull request fixes the issue:
Adds additional logging to know what windows could be focused and which currently have focus. Don't focus a window if many windows match our targeted window filter.