Fix dirty files on translation system
The msg path is filtered in the translation system (via git config). It removes the file path
from the comments of *.po files.
Recently git has been complaining about files that have not yet had these lines removed.
It claims the files are modified, and will not checkout a file that conflicts with it.
This is despite the working directory matching HEAD, git still thinks there is a diff.