Remove the one and only use of ngettext as our po checks seem to be getting confused by it, and this was not the first and only issue. (#12448)
Recently NVDA began using ngettext to allow translators to provide both singular and plural translations for a particular message.
However, it seems that our PO validation code in the translation system is not handling plural forms well, and is incorrectly classing translations as missing brace format variables, as it is incorrectly comparing the translation with the message before that one.
It then outputs the following error:
Message starting on line 7169
Original: "{startTime} to {endTime}"
Translated: "category {categories}"
Error: missing brace format interpolation, extra brace format interpolation
Expected: these brace format interpolations: {endTime}, {startTime}
Got: these brace format interpolations: {categories} 
ngettext was so far only used for one string in NVDA, in source/appModules/outlook.py. Seen by doing git grep ngettext.
We have many many strings in NVDA which would benefit from ngettext, but we need to know it works correctly. So until we are sure we can validate this correctly, we should stop using it in the one and only place it exists today.

Description of how this pull request fixes the issue:
Stop using ngettext for the one string in the Outlook appModule, instead replacing it with a normal use of _() providing just the plural form.