Add typing for virtual buffer / text info / fields normalaziation
String type can also be returned from XMLFormatting.XMLTextParser.parse
Update typing to include optional string.
Since data could be an optional string, but it is apparently not ommited
when None, log a warning. This will help us to track if it ever is None.

If it is not ever None, explicitly exclude None and tighten up the
typing.