Ensure NVDA quits from menu by queueing exit process (#12499)
When the exit dialog is disabled, the popup menu stays activated while exiting. The NVDA exit process creates a call to destroy the popup menu while it is still activated. This causes a wx issue.

Description of how this pull request fixes the issue:
Always queue the exit process. This allows the PopupMenu to close before attempting to be deleted.