Clarify log message for trigger NVDA exit
There is only an error case when GUI intiated exiting is not first
to trigger the exit.

Previously there was a warning for every exit, since there is
always a "catch-all" exit trigger.
The "catch-all" exit trigger is necessary to ensure there cleanup happens
when a WM_QUIT windows message is recieved, causing wx to exit the
MainLoop.
Now there is only a message when the "catch-all" is first, aimed at helping
developers identify when exit was due to WM_QUIT.