Force a new NVDA instance to start after crashing (#12562)
When NVDA crashes, core.restart is called. #12548 changed the behaviour of core.restart to safely shutdown NVDA and queue a new instance to start. NVDA cannot exit safely or queue the new instance to start if the thread is in an uncertain state (eg crashed, frozen).

This commit makes NVDA directly start a new instance of NVDA if a crash occurs.