nvda_slave should log at debug and parent NVDA can filter (#12591)
For some time now, nvda_slave has not been logging back to NVDA, even if NVDA was set to log level debug. E.g. the nvda_install.log appveyor artifact does not contain any installer messages. Similarly, the COM registration fixing tool also looks like it logs nothing unless there is a specific error.
When nvda_slave initializes logging, it sets the level of loghandler.log to 0. I'm guessing that we chose 0 as it was thought to be the lowest (most detailed) level (I.e. greater than debug (10)). However, 0 seems to have a special meaning of NOTSET, and therefore messages honor the first ancestor that is not set to NOTSET, which is the root logger, which is probably set to warning by default. Thus, all the debug messages are being dropped and not being sent back to NVDA, even if NVDA's own log level was debug.
Description of how this pull request fixes the issue:
In nvda_slave: initialize logging with a specific level of DEBUG (10), the most detailed level.