Check that a remote update exists even if a pending update is cached (#12592)
if a pending update is cached, even if there is no remote update, we attempt to compare the versions. This seems like an uncommon bug, possibly brought to surface by earlier beta updates failing.

Description of how this pull request fixes the issue:
Added a trail of typing to confirm the possibility of a state where there is a pending update cached and no remote update. Added a check to ensure that both of these exist before comparing their versions.