Create separate AppVeyor projects (#12615)
If many PR/try builds are queued, it can take a long time to build a release. Building tagged releases is an involved process for an NV Access developer and as such, should be prioritised.
This commit creates two separate new projects for tagged releases and other builds. This allows for priorities to be set in the online UI.