Exit via WM_QUIT properly (#12616)
starting a new instance kills the existing process via WM_QUIT. This exits NVDA by exiting the main loop without exiting NVDA properly. We currently add an event to shutdown properly to the event queue, but the shutdown continues without waiting for that to complete.

Description of how this pull request fixes the issue:
This pumps the event queue before continuing if exiting via WM_QUIT to ensure the shutdown happens properly.