Check system tests for crashes, add test for auxiliary thread crashes (#12619)
System tests may pass even if a crash occurs during them.
Similar to #12604, it is important to ensure NVDA can restart when an auxiliary thread crashes.

This adds a check for a crash dump file when we save the NVDA log, tests will fail if a crash has occurred during the test. Crash dump files will be saved like log files.

Adds system tests to simulate the UIAHandler and braille thread crashing.