Protect the ssh key with stronger encryption, similar to the authenticode certificate. (#12635)
Appveyor contains an ssh private key that allows it to log into an NV Access server to deploy snapshots / releases. The current ssh key in the repository, although encrypted with an appveyor secret variable, was not as strong as it could have been, and during decryption, in the appveyor build, openssl would warn about this.

This pr replaces the ssh key with a newly generated one, encrypted using a new appveyor secret variable, and the encryption of the key is also significantly stronger
as it uses SHA256 instead of md5, a salt is now used, and a newer secret key derivation algorithm is now used, as recommended by openssl.
This ssh key is now encrypted just as strongly as the authenticode certificate also in this repository.