Vocalize paragraph indent and hanging indent shortcuts in MS Word (#10415)
In MS Word, the result of indent and hanging indent shortcuts is not announced by NVDA. It may be useful for the user to know the distance from the left edge of the page, as done when pressing Tab or Shift+Tab.

Description of how this pull request fixes the issue:
When the following shortcuts are executed, the distance from the left edge of the page of the paragraph is reported:

Ctrl+M: indent
Ctrl+Shift+M: outdent
Ctrl+T: hanging indent
Ctrl+Shift+T: hanging outdent
The reported distance is the one from left edge of page to the beginning of all line of the paragraph except the first that may differ.

The listed shortcuts are the ones of English version of MS Word. They may differ for other languages. In this case, translators will need to update gesture.ini file accordingly.

Co-authored-by: buddsean <sean@nvaccess.org>