UI Automation in Windows Console: add "API levels" as a better isImprovedTextRangeAvailable (#12660)
Summary of the issue:
The Windows 11 inbox console will ship with most, but not all, fixes required to make UIA by default a good experience. However, a new conhost version will be delivered "out-of-box" including additional functionality and bug fixes. Therefore, isImprovedTextRangeAvailable (a boolean value) is now insufficient to express the range of possible UIA console versions now in the wild.

Description of how this pull request fixes the issue:
Adds an enumeration, WinConsoleAPILevel, with comments describing the meaning of each member.
Refactors NVDA's current code to use the new enum without changing current functionality. Functionality changes to follow in subsequent PRs.
Expose apiLevel in dev info to ease debugging.
Deprecate isImprovedTextRangeAvailable, but refactor it for now to use apiLevel.

Co-authored-by: Leonard de Ruijter <leonardder@users.noreply.github.com>
Co-authored-by: buddsean <sean@nvaccess.org>