Input Gesture Dialog fix for when there is no filtered Input Gestures (#12693)
Fixes #12673

Summary of the issue:
The input gestures GUI did not handle the case where there was no filtered input gestures available, such as filtering the gestures for a string like "zzzzzzzzz". This results in multiple input gesture dialogs being able to be opened at once.

Add checks for when the filtered gesture list is empty.