Expand checkPot to warn about messages which were marked as not having translator comments but are no longer present in the translations template (#12705)
Summary of the issue:
Currently checkPot has a list of messages which are translatable but don't have commends for translators. While it warns about messages in this list for which comments are added it does not warn about these which are no longer present in the .pot template.

Description of how this pull request fixes the issue:
checkPot was modified so that if there are messages in EXPECTED_MESSAGES_WITHOUT_COMMENTS which were removed from the source code it lists them and fails similarly to the case in which a new message has no comments for translators.